data-warehouse/src/requirements.txt
distlib==0.3.6
filelock==3.12.0
pip==21.3.1
platformdirs==3.5.1
py4j==0.10.9.7
pyspark==3.4.0
python-dotenv==1.0.0
schedule==1.2.0
setuptools==65.7.0
virtualenv==20.23.0
